(svn r19393) [1.0] -Backport from trunk: - Fix: One could not level the whole map anymore at once (r19392) - Fix: Only show the 'No AIs available' error message when explicitly changing the number of AI opponents [FS3676] (r19389) - Fix: [NoAI] When reloading a savegame, an AI failing to compile could trigger (trying) to read the not yet loaded information of another AI via the AI Debug window and its "open with the most recently used AI" feature [FS#3666] (r19388) - Fix: Close all orders windows when switching companies [FS#3671] (r19387) - Fix: [IPv6] Netmask calculations were wrong if cidr >= 32 [FS#3684] (r19385) - Fix: Overbuilding bridges, rail stations did not properly update PBS reservation [FS#3680] (r19384, r19383)

+++ src/tilearea_type.h
@@ -17,8 +17,8 @@
 /** Represents the covered area of e.g. a rail station */
 struct TileArea {
 	TileIndex tile; ///< The base tile of the area
-	uint8 w;        ///< The width of the area
-	uint8 h;        ///< The height of the area
+	uint16 w;       ///< The width of the area
+	uint16 h;       ///< The height of the area
 
 	/** Just construct this tile area */
 	TileArea() {}
